Només falta el video.

+++ Laia Rodríguez Ramos/index.php
@@ -16,8 +16,10 @@
     <?php
         if (!isset($_COOKIE[$cookieLang])) {
             //echo 'La cookie es la default. CAT.';
-            setcookie($cookieLang, $cookieCat, $tempsDuracio);
-            cookieValueCat();
+            //setcookie($cookieLang, $cookieCat, $tempsDuracio);
+            //AQUI AL FER EL SETCOOKIE, ES FEIA EL BUCLE AMB DELETE I PER AIXO
+            //LA COOKIE NO S'ESBORRABA MAI.
+            cookieValueDefault();
         } else if(isset($_COOKIE[$cookieLang])){
             if($_COOKIE[$cookieLang] == "es"){
                 cookieValueEs();
@@ -30,17 +32,17 @@
             }
         }
 
-        //AIXO HAURIA DE SER UNA FUNCIO PER CATALÀ
-        function cookieValueCat(){
+        function cookieValueDefault(){
             if (isset($_SESSION["LoggedIn"]) && $_SESSION["LoggedIn"] === true) {
                 //SI ES ALUMNE
                 if ($_SESSION['rol'] === 'Alumne') {
 
                     echo "<h1>Benvingut, " . $_SESSION["username"] . "! Ets un: " . $_SESSION["rol"] . "</h1>";?>
-                    
+  
                     <a href="idioma.php?idioma=cat">CAT</a>
                     <a href="idioma.php?idioma=es">ES</a>
                     <a href="idioma.php?idioma=en">EN</a>
+
                     <a href="delete.php">Eliminar</a>
                     <br><br>
                     <a href="mostrarInfo.php?id=<?php echo $_SESSION["user_id"]; ?> ">Mostrar informacio</a>
@@ -80,9 +82,80 @@
                     }
 
                     ?>
+
                     <a href="idioma.php?idioma=cat">CAT</a>
                     <a href="idioma.php?idioma=es">ES</a>
                     <a href="idioma.php?idioma=en">EN</a>
+                    
+                    <a href="delete.php">Eliminar</a>
+                    <br><br>
+                    <a href="mostrarInfo.php?id=<?php echo $_SESSION["user_id"]; ?> ">Mostrar informacio</a>
+                    <a href="desconnectar.php">Desconnectar</a>
+                    
+                    <?php
+                }
+            }
+            else{
+                header('Location: logUser.html');
+            }
+        }
+
+        //AIXO HAURIA DE SER UNA FUNCIO PER CATALÀ
+        function cookieValueCat(){
+            if (isset($_SESSION["LoggedIn"]) && $_SESSION["LoggedIn"] === true) {
+                //SI ES ALUMNE
+                if ($_SESSION['rol'] === 'Alumne') {
+
+                    echo "<h1>Benvingut, " . $_SESSION["username"] . "! Ets un: " . $_SESSION["rol"] . "</h1>";?>
+  
+                    <a href="idioma.php?idioma=cat" style="<?php echo ($_COOKIE['lang'] === 'cat') ? ' color: red;' : ' color: blue;'; ?>">CAT</a>
+                    <a href="idioma.php?idioma=es" style="<?php echo ($_COOKIE['lang'] === 'es') ? ' color: red;' : ' color: blue;'; ?>">ES</a>
+                    <a href="idioma.php?idioma=en" style="<?php echo ($_COOKIE['lang'] === 'en') ? ' color: red;' : ' color: blue;'; ?>">EN</a>
+
+                    <a href="delete.php">Eliminar</a>
+                    <br><br>
+                    <a href="mostrarInfo.php?id=<?php echo $_SESSION["user_id"]; ?> ">Mostrar informacio</a>
+                    <a href="desconnectar.php">Desconnectar</a>
+                    
+                    <?php
+
+                } 
+                //SI ES PROFESSOR
+                elseif ($_SESSION['rol'] === 'Professor') {
+
+                    echo "<h1>Benvingut, " . $_SESSION["username"] . "! Ets un: " . $_SESSION["rol"] . "</h1>";
+
+                    $query = "SELECT username, surname, email FROM userlaia";
+
+                    $connect = mysqli_connect(DB_HOST,DB_USER,DB_PSW,DB_NAME);
+                    $result = mysqli_query($connect, $query);
+
+                    echo "<h2>Llista d'usuaris:</h2>";
+
+                    //UN ALTRE COP, SI HI HA UN USUARI COM A MINIM, GUARDA AL ARRAY
+                    if ($result && mysqli_num_rows($result) > 0) {
+                        $usuaris = array();
+                        
+                        //RECORRER LA BBDD USERSLAIA
+                        while ($row = mysqli_fetch_assoc($result)) {
+                            $usuaris[] = $row;
+                        }
+                        
+                        //GUARDA NOM, COGNOM A L'ARRAY PER LLAVORS IMPRIMIR-HO
+                        foreach ($usuaris as $usuari) {
+                            $username = $usuari['username'];
+                            $surname = $usuari['surname'];
+                            $email = $usuari['email'];
+                            echo "Nom i cognom: $username $surname <br> Correu: $email <br><br>";
+                        }
+                    }
+
+                    ?>
+
+                    <a href="idioma.php?idioma=cat" style="<?php echo ($_COOKIE['lang'] === 'cat') ? ' color: red;' : ' color: blue;'; ?>">CAT</a>
+                    <a href="idioma.php?idioma=es" style="<?php echo ($_COOKIE['lang'] === 'es') ? ' color: red;' : ' color: blue;'; ?>">ES</a>
+                    <a href="idioma.php?idioma=en" style="<?php echo ($_COOKIE['lang'] === 'en') ? ' color: red;' : ' color: blue;'; ?>">EN</a>
+                    
                     <a href="delete.php">Eliminar</a>
                     <br><br>
                     <a href="mostrarInfo.php?id=<?php echo $_SESSION["user_id"]; ?> ">Mostrar informacio</a>
@@ -104,9 +177,10 @@
 
                     echo "<h1>Bienvenido, " . $_SESSION["username"] . "! Eres un: " . $_SESSION["rol"] . "</h1>";?>
                     
-                    <a href="idioma.php?idioma=cat">CAT</a>
-                    <a href="idioma.php?idioma=es">ES</a>
-                    <a href="idioma.php?idioma=en">EN</a>
+                    <a href="idioma.php?idioma=cat" style="<?php echo ($_COOKIE['lang'] === 'cat') ? ' color: red;' : ' color: blue;'; ?>">CAT</a>
+                    <a href="idioma.php?idioma=es" style="<?php echo ($_COOKIE['lang'] === 'es') ? ' color: red;' : ' color: blue;'; ?>">ES</a>
+                    <a href="idioma.php?idioma=en" style="<?php echo ($_COOKIE['lang'] === 'en') ? ' color: red;' : ' color: blue;'; ?>">EN</a>
+                    
                     <a href="delete.php">Eliminar</a>
                     <br><br>
                     <a href="mostrarInfo.php?id=<?php echo $_SESSION["user_id"]; ?> ">Mostrar información</a>
@@ -146,9 +220,11 @@
                     }
 
                     ?>
-                    <a href="idioma.php?idioma=cat">CAT</a>
-                    <a href="idioma.php?idioma=es">ES</a>
-                    <a href="idioma.php?idioma=en">EN</a>
+
+                    <a href="idioma.php?idioma=cat" style="<?php echo ($_COOKIE['lang'] === 'cat') ? ' color: red;' : ' color: blue;'; ?>">CAT</a>
+                    <a href="idioma.php?idioma=es" style="<?php echo ($_COOKIE['lang'] === 'es') ? ' color: red;' : ' color: blue;'; ?>">ES</a>
+                    <a href="idioma.php?idioma=en" style="<?php echo ($_COOKIE['lang'] === 'en') ? ' color: red;' : ' color: blue;'; ?>">EN</a>
+                    
                     <a href="delete.php">Eliminar</a>
                     <br><br>
                     <a href="mostrarInfo.php?id=<?php echo $_SESSION["user_id"]; ?> ">Mostrar información</a>
@@ -169,9 +245,10 @@
 
                     echo "<h1>Welcome, " . $_SESSION["username"] . "! You are: " . $_SESSION["rol"] . "</h1>";?>
                     
-                    <a href="idioma.php?idioma=cat">CAT</a>
-                    <a href="idioma.php?idioma=es">ES</a>
-                    <a href="idioma.php?idioma=en">EN</a>
+                    <a href="idioma.php?idioma=cat" style="<?php echo ($_COOKIE['lang'] === 'cat') ? ' color: red;' : ' color: blue;'; ?>">CAT</a>
+                    <a href="idioma.php?idioma=es" style="<?php echo ($_COOKIE['lang'] === 'es') ? ' color: red;' : ' color: blue;'; ?>">ES</a>
+                    <a href="idioma.php?idioma=en" style="<?php echo ($_COOKIE['lang'] === 'en') ? ' color: red;' : ' color: blue;'; ?>">EN</a>
+                    
                     <a href="delete.php">Delete</a>
                     <br><br>
                     <a href="mostrarInfo.php?id=<?php echo $_SESSION["user_id"]; ?> ">Show Data</a>
@@ -211,9 +288,11 @@
                     }
 
                     ?>
-                    <a href="idioma.php?idioma=cat">CAT</a>
-                    <a href="idioma.php?idioma=es">ES</a>
-                    <a href="idioma.php?idioma=en">EN</a>
+
+                    <a href="idioma.php?idioma=cat" style="<?php echo ($_COOKIE['lang'] === 'cat') ? ' color: red;' : ' color: blue;'; ?>">CAT</a>
+                    <a href="idioma.php?idioma=es" style="<?php echo ($_COOKIE['lang'] === 'es') ? ' color: red;' : ' color: blue;'; ?>">ES</a>
+                    <a href="idioma.php?idioma=en" style="<?php echo ($_COOKIE['lang'] === 'en') ? ' color: red;' : ' color: blue;'; ?>">EN</a>
+                    
                     <a href="delete.php">Delete</a>
                     <br><br>
                     <a href="mostrarInfo.php?id=<?php echo $_SESSION["user_id"]; ?> ">Show data</a>
